Improve mesh preprocessing script

+++ apps/mesh_preprocessing/mesh_preprocessing.sh
@@ -46,7 +46,6 @@ case ${answer:0:1} in
         read -rp "Triangle budget per KD-tree node: " KDTREE_TRI_BUDGET
         read -rp "Chart creation cost threshold: " COST_THRESHOLD
         read -rp "Triangle budget per BVH node: " TRI_BUDGET
-        read -rp "Maximum size of final texture: " MAX_FINAL_TEX_SIZE
         read -rp "Number of dilations: " NUM_DILATIONS
     ;;
 esac
@@ -69,10 +68,8 @@ read -p "Convert all textures to PNG (required)? (y/n)? " answer
 case ${answer:0:1} in
     y|Y )
         # convert textures to PNG using mogrify
-        targets=`*.jpg *.JPG *.png *.PNG *.bmp *.BMP *.tiff *.TIFF *.tif *.TIF *.ppm *.PPM *.pgm *.PGM *.pbm *.PBM *.pnm *.PNM`
-        echo "Converting to PNG: "
-        echo ${targets}
-        mogrify -format png ${targets}
+        echo "Converting to PNG... "
+        mogrify -format png *.jpg
     ;;
     * )
         echo "Skipping conversion (assuming was done before)"
@@ -82,24 +79,31 @@ esac
 read -p "Flip all textures (required)? (y/n)? " answer
 case ${answer:0:1} in
     y|Y )
-        targets=`*.png`
-        echo "Flipping PNGs"
-        echo ${targets}
-        mogrify -flip ${targets}
+        echo "Flipping PNGs..."
+        mogrify -flip *.jpg
     ;;
     * )
         echo "Skipping flipping (assuming was done before)"
     ;;
 esac
 
+read -rp "Please input the maximum allowed size for the final texture: " MAX_FINAL_TEX_SIZE
+
 echo -e "\e[0m"
 
-echo "Running chart creation with file $SRC_OBJ"
-echo "-----------------------------------------"
-
 DATE=`date '+%Y-%m-%d:%H:%M:%S'`
-time ${LAMURE_DIR}lamure_mesh_preprocessing -f ${OBJPATH} -tkd ${KDTREE_TRI_BUDGET} -co ${COST_THRESHOLD} -tbvh ${TRI_BUDGET} -multi-max ${MAX_FINAL_TEX_SIZE} 2>&1 | tee log_${DATE}.txt
 
+read -p "Would you like to produce a .raw file for VT preprocessing at the end (if not, you get a .png)? (y/n)? " answer
+case ${answer:0:1} in
+    y|Y )
+        echo "Going to create RAW texture file at the end."
+        time ${LAMURE_DIR}lamure_mesh_preprocessing -f ${OBJPATH} -raw -tkd ${KDTREE_TRI_BUDGET} -co ${COST_THRESHOLD} -tbvh ${TRI_BUDGET} -multi-max ${MAX_FINAL_TEX_SIZE} 2>&1 | tee log_${DATE}.txt
+    ;;
+    * )
+        echo "Going to create PNG texture file at the end."
+        time ${LAMURE_DIR}lamure_mesh_preprocessing -f ${OBJPATH} -tkd ${KDTREE_TRI_BUDGET} -co ${COST_THRESHOLD} -tbvh ${TRI_BUDGET} -multi-max ${MAX_FINAL_TEX_SIZE} 2>&1 | tee log_${DATE}.txt
+    ;;
+esac
